+++ examples/virtual-field/src/primatives/index.ts
@@ -1,0 +1,12 @@
+import { AutocompleteSelect } from "./autocomplete-select"
+import { OrderableList } from "./orderable"
+
+export type Item = {
+    label: string;
+    value?: string;
+  }
+export type OrderableItem = Item & {
+    key: string;
+  }
+
+export { AutocompleteSelect, OrderableList }
